refactor: simplify authentication components and rename navbar directories
