+++ apps/nextjs/src/app/(main)/(clerk-auth)/sign-up/[[...sign-up]]/page.tsx
@@ -1,16 +1,14 @@
 import { Suspense } from "react";
-import { ClerkProvider, SignedOut, SignUp } from "@clerk/nextjs";
+import { SignedOut, SignUp } from "@clerk/nextjs";
 
 export default function SignUpPage() {
   return (
-    <ClerkProvider dynamic>
-      <SignedOut>
-        <div className="flex h-full min-h-screen w-full items-center justify-center">
-          <Suspense fallback={<div>Loading...</div>}>
-            <SignUp />
-          </Suspense>
-        </div>
-      </SignedOut>
-    </ClerkProvider>
+    <SignedOut>
+      <div className="flex h-full min-h-screen w-full items-center justify-center">
+        <Suspense fallback={<div>Loading...</div>}>
+          <SignUp />
+        </Suspense>
+      </div>
+    </SignedOut>
   );
 }
