+++ apps/nextjs/src/app/(main)/@navbar/_components/navbar.tsx
@@ -1,6 +1,6 @@
-import NavbarContainer from "~/app/(main)/@navbar/_navbar-components/navbar-container";
-import NavbarLinks from "~/app/(main)/@navbar/_navbar-components/navbar-links";
-import RightMenu from "~/app/(main)/@navbar/_navbar-components/navbar-right-menu";
+import NavbarContainer from "~/app/(main)/@navbar/_components/navbar-container";
+import NavbarLinks from "~/app/(main)/@navbar/_components/navbar-links";
+import RightMenu from "~/app/(main)/@navbar/_components/navbar-right-menu";
 import BattleStadium from "~/components/battle-stadium";
 
 export default function Navbar() {
