+++ apps/nextjs/src/app/(main)/@navbar/default.tsx
@@ -1,4 +1,4 @@
-import Navbar from "~/app/(main)/@navbar/_navbar-components/navbar";
+import Navbar from "~/app/(main)/@navbar/_components/navbar";
 
 export default function NavbarDefaultSlot() {
   return <Navbar />;
